solved

+++ exercises/02_basic_calculator/01_integers/src/lib.rs
@@ -1,4 +1,4 @@
-fn compute(a: u32, b: u32) -> u32 {
+fn compute(a: u8, b: u8) -> u8 {
     // TODO: change the line below to fix the compiler error and make the tests pass.
     let multiplier: u8 = 4;
     a + b * multiplier
